Qt_GUI/requirements.txt
GitPython
socket
paho-mqtt
